BinaryVectorByClass base version added decimal changed data type changed to double

+++ Pierniczek/Models/RangeModel.cs
@@ -13,7 +13,7 @@ namespace Pierniczek.Models
         {
         }
 
-        public decimal? Min { get; set; }
-        public decimal? Max { get; set; }
+        public double? Min { get; set; }
+        public double? Max { get; set; }
     }
 }
